Added gamecube support for byte sized packets sent by the NicoHood Nintendo API
This adds support to reads packets in the form of bytes that are sent from the NicoHood Nintendo API at https://github.com/NicoHood/Nintendo

The reason for this is that when using the above API, it's incredibly time sensitive to write data that is read from the controller to the console, so to save time we just call a single write for each packet (and write the whole gamecube report) via Serial.write(report.raw8, sizeof(report.raw8)).

It however is in a different format, thus the need for the NICOHOOD_BUTTONS setup.

Currently this only supports gamecube input but will be extended to N64 in the near future.